requirements.txt
# Core dependencies with versions matching setup_env.sh
torch==2.0.0
torchvision==0.15.0
torchaudio==2.0.0

# Data processing and analysis
numpy<2.0.0
pandas>=1.3.0
scikit-learn>=0.24.2
pyarrow>=6.0.0

# Deep learning
transformers>=4.5.0
datasets>=2.0.0
sentencepiece>=0.1.96

# Bioinformatics
biopython>=1.79
regex>=2022.1.18

# Visualization
matplotlib>=3.4.3
ipywidgets>=7.6.5

# Jupyter environment
jupyterlab>=3.2.8
notebook>=6.4.8
tqdm>=4.62.3

# Configuration and logging
pyyaml>=5.4.1
geoparse>=1.0.0

# Development tools (as used in setup_env.sh)
ipython>=8.0.1
jupyter>=1.0.0

# For RNA structure analysis (if needed)
ViennaRNA>=2.5.0

# Ensure compatibility with DNABERT-2
# Additional dependencies from DNABERT_2/requirements.txt will be installed
# during setup_env.sh execution